[MT-5235] Extract `EntryPoint` component into a separate file
